requirements.txt
python-telegram-bot
python-telegram-bot[socks]
requests